requirements.txt

tox>=3.18.0
twine==3.2.0
pytest>=5.4.3
flake8>=3.8.3
dbus-python==1.2.18
check-manifest>=0.42